+++ app/code/local/Radarsofthouse/Reepay/Helper/Data.php
@@ -492,22 +492,24 @@ class Radarsofthouse_Reepay_Helper_Data extends Mage_Core_Helper_Abstract
             $_paymentMethods[] = 'resurs';
         } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_forbrugsforeningen') {
             $_paymentMethods[] = 'ffk';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_vipps') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_vipps') {
             $_paymentMethods[] = 'vipps';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_googlepay') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_googlepay') {
             $_paymentMethods[] = 'googlepay';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_ideal') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_ideal') {
             $_paymentMethods[] = 'ideal';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_blik') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_blik') {
             $_paymentMethods[] = 'blik';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_p24') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_p24') {
             $_paymentMethods[] = 'p24';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_verkkopankki') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_verkkopankki') {
             $_paymentMethods[] = 'verkkopankki';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_giropay') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_giropay') {
             $_paymentMethods[] = 'giropay';
-        }elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_sepa') {
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_sepa') {
             $_paymentMethods[] = 'sepa';
+        } elseif ($order->getPayment()->getMethodInstance()->getCode() == 'reepay_bancontact') {
+            $_paymentMethods[] = 'bancontact';
         } else {
             $paymentMethods = $this->getConfig('allowwed_payment');
             $_paymentMethods = explode(',', $paymentMethods);