+++ app/code/local/Radarsofthouse/Reepay/Model/System/Config/Source/Dropdown/Allowwedpayment.php
@@ -44,9 +44,7 @@ class Radarsofthouse_Reepay_Model_System_Config_Source_Dropdown_Allowwedpayment
             array('value' => 'ideal', 'label' => __('iDEAL')),
             array('value' => 'blik', 'label' => __('BLIK')),
             array('value' => 'p24', 'label' => __('Przelewy24 (P24)')),
-            array('value' => 'verkkopankki', 'label' => __('Verkkopankki')),
-            array('value' => 'giropay', 'label' => __('giropay')),
-            array('value' => 'sepa', 'label' => __('SEPA Direct Debit'))
+            array('value' => 'giropay', 'label' => __('giropay'))
         );
     }
 }